streamlit/pyproject.toml
[project]
name = "imessage-dashboard"
version = "0.1.0"
description = "iMessage Search Dashboard"
authors = [
    {name = "Shrey"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "qdrant-client (>=1.16.1,<2.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
